big refactor of all things related to build:
- updated babel to version 7
- updated webpack to version 4
- now making 3 distinct builds of Klaro:
	- one "umd" build that includes klaro and all its dependencies
in order to work directly when including the script in the browser. It's
the same build as before, except now it includes babel-runtime to make
sure it works on the listed browsers (especially ie11)
	*note*: it actually doesn't work on ie11 yet, because of the
usage of document.currentScript. We'll ditch this soon
	- one "cjs" build which is just the source code transpiled with
babel in a commonjs env
	- one "es" build which is just the source code transpiled with
babel in a es modules env
- removed the klaro.version() function, I guess I could add klaro
version in a comment at the top of the build instead
- removed the webpack dev server configuration as it's not that useful
and it was taking quite a bit of code in the config
- better handling of dependencies vs devDependencies so that everything
works when consuming Klaro as an npm package
- tracking the package-lock in git