Removed CSS namespacing option
Another option would have been to make every class name fully
configurable (the whole class name instead of just the prefix), but the
cost of it would outweigh the gain.
Also, the `stylePrefix` option wasn't even documented, so we're betting
on the fact that very few people are actually using it, if any.