Fixed type declarations for translations
`tsup` was generating a custom type for each and every translation
object, with all of their properties, resulting in a huge declaration.
It turns out that `satisfies` does not impact the output, whereas `as`
forces the `Translation` type to be used.
We're still keeping the `satisfies` part, as it provides better
typechecking (i.e. warning about missing or unknown properties).